try-nocho-json-converter-v8

+++ app/src/main/java/com/example/forecastmvvm/data/db/entity/Current/Clouds.kt
@@ -1,5 +1,5 @@
 package com.example.forecastmvvm.data.db.entity.Current
 
 data class Clouds(
-    val all: Int // 40
+    val all: Double // 40
 )
