+++ app/src/main/java/com/example/forecastmvvm/data/db/entity/Current/Wind.kt
@@ -2,5 +2,5 @@ package com.example.forecastmvvm.data.db.entity.Current
 
 data class Wind(
     val speed: Double, // 3.6
-    val deg: Int // 160
+    val deg: Double // 160
 )
